test: extract a shared test harness
Move the copy-pasted test runner (test, assert, assertEqual, summary)
into src/test/harness.ts. Both calculate.test.ts and
versionedStore.test.ts now import from it. The versionedStore wrapper
keeps its resetStorage() pre-hook locally. Exclude src/test/** from
tsconfig so the harness (which calls process.exit) isn't type-checked
against DOM-only libs, consistent with how the .test.ts files are
already excluded.
